Simplify final release
