+++ extra/update-version.js
@@ -5,10 +5,8 @@ const util = require("../src/util");
 
 util.polyfill();
 
-const oldVersion = pkg.version;
-const newVersion = process.argv[2];
+const newVersion = process.env.VERSION;
 
-console.log("Old Version: " + oldVersion);
 console.log("New Version: " + newVersion);
 
 if (! newVersion) {
@@ -22,16 +20,13 @@ if (! exists) {
 
     // Process package.json
     pkg.version = newVersion;
-    pkg.scripts.setup = pkg.scripts.setup.replaceAll(oldVersion, newVersion);
-    pkg.scripts["build-docker"] = pkg.scripts["build-docker"].replaceAll(oldVersion, newVersion);
-    pkg.scripts["build-docker-alpine"] = pkg.scripts["build-docker-alpine"].replaceAll(oldVersion, newVersion);
-    pkg.scripts["build-docker-debian"] = pkg.scripts["build-docker-debian"].replaceAll(oldVersion, newVersion);
+
+    // Replace the version: https://regex101.com/r/hmj2Bc/1
+    pkg.scripts.setup = pkg.scripts.setup.replace(/(git checkout )([^\s]+)/, `$1${newVersion}`);
     fs.writeFileSync("package.json", JSON.stringify(pkg, null, 4) + "\n");
 
     commit(newVersion);
     tag(newVersion);
-
-    updateWiki(oldVersion, newVersion);
 
 } else {
     console.log("version exists");
@@ -64,37 +59,3 @@ function tagExists(version) {
     return res.stdout.toString().trim() === version;
 }
 
-function updateWiki(oldVersion, newVersion) {
-    const wikiDir = "./tmp/wiki";
-    const howToUpdateFilename = "./tmp/wiki/🆙-How-to-Update.md";
-
-    safeDelete(wikiDir);
-
-    child_process.spawnSync("git", ["clone", "https://github.com/louislam/uptime-kuma.wiki.git", wikiDir]);
-    let content = fs.readFileSync(howToUpdateFilename).toString();
-    content = content.replaceAll(`git checkout ${oldVersion}`, `git checkout ${newVersion}`);
-    fs.writeFileSync(howToUpdateFilename, content);
-
-    child_process.spawnSync("git", ["add", "-A"], {
-        cwd: wikiDir,
-    });
-
-    child_process.spawnSync("git", ["commit", "-m", `Update to ${newVersion} from ${oldVersion}`], {
-        cwd: wikiDir,
-    });
-
-    console.log("Pushing to Github");
-    child_process.spawnSync("git", ["push"], {
-        cwd: wikiDir,
-    });
-
-    safeDelete(wikiDir);
-}
-
-function safeDelete(dir) {
-    if (fs.existsSync(dir)) {
-        fs.rmdirSync(dir, {
-            recursive: true,
-        });
-    }
-}
